move the networks to kale.embed
