+++ examples/materials_benchmark/configs/finetune/leftnet/leftnet_prop.yaml
@@ -1,5 +1,5 @@
 DATASET:
- TRAIN: 'data/fine_tune/train_data.json'
+ TRAIN: 'data/fine_tune/test_data.json'
  VAL: 'data/fine_tune/test_data.json'
 
 MODEL:
