--- examples/materials_benchmark/models/leftnet.py
@@ -1,530 +1,0 @@
-# The following code is a modified version of the original code from the LEFTNet repository (https://github.com/yuanqidu/LeftNet)
-
-import math
-from math import pi
-from typing import Optional, Tuple
-
-import torch
-from torch import nn
-from torch.nn import Embedding
-from torch_geometric.nn import radius_graph
-from torch_geometric.nn.conv import MessagePassing
-from torch_scatter import scatter
-import torch.nn.functional as F
-
-class rbf_emb(nn.Module):
-    """
-    modified: delete cutoff with r
-    """
-
-    def __init__(self, num_rbf, rbound_upper, rbf_trainable=False):
-        super().__init__()
-        self.rbound_upper = rbound_upper
-        self.rbound_lower = 0
-        self.num_rbf = num_rbf
-        self.rbf_trainable = rbf_trainable
-        means, betas = self._initial_params()
-
-        self.register_buffer("means", means)
-        self.register_buffer("betas", betas)
-
-    def _initial_params(self):
-        start_value = torch.exp(torch.scalar_tensor(-self.rbound_upper))
-        end_value = torch.exp(torch.scalar_tensor(-self.rbound_lower))
-        means = torch.linspace(start_value, end_value, self.num_rbf)
-        betas = torch.tensor([(2 / self.num_rbf * (end_value - start_value)) ** -2] * self.num_rbf)
-        return means, betas
-
-    def reset_parameters(self):
-        means, betas = self._initial_params()
-        self.means.data.copy_(means)
-        self.betas.data.copy_(betas)
-
-    def forward(self, dist):
-        dist = dist.unsqueeze(-1)
-        rbounds = 0.5 * (torch.cos(dist * pi / self.rbound_upper) + 1.0)
-        rbounds = rbounds * (dist < self.rbound_upper).float()
-        return rbounds * torch.exp(-self.betas * torch.square((torch.exp(-dist) - self.means)))
-
-
-class NeighborEmb(MessagePassing):
-    def __init__(self, hid_dim: int, input_dim: int):
-        super(NeighborEmb, self).__init__(aggr="add")
-        self.hid_dim = hid_dim
-        self.input_dim = input_dim
-        self.fc = nn.Linear(self.input_dim, self.hid_dim)
-        self.ln_emb = nn.LayerNorm(hid_dim, elementwise_affine=False)
-
-    def forward(self, z, edge_index, embs):
- 
-        """Atom Embedding + Neighborhours Embedding"""
-        z_emb = self.ln_emb(self.fc(z)) # shape: (num_nodes, hidden_channels)
-        s_neighbors = self.propagate(edge_index, x=z_emb, norm=embs)
-        z_emb = z_emb + s_neighbors
-        # s = s + s_neighbors
-        return z_emb # shape: (num_nodes, hidden_channels)
-
-    def message(self, x_j, norm):
-        return norm.view(-1, self.hid_dim) * x_j
-
-
-class S_vector(MessagePassing):
-    def __init__(self, hid_dim: int):
-        super(S_vector, self).__init__(aggr="add")
-        self.hid_dim = hid_dim
-        self.lin1 = nn.Sequential(
-            nn.Linear(hid_dim, hid_dim), nn.LayerNorm(hid_dim, elementwise_affine=False), nn.SiLU()
-        )
-
-    def forward(self, s, v, edge_index, emb):
-        s = self.lin1(s)
-        emb = emb.unsqueeze(1) * v
-
-        v = self.propagate(edge_index, x=s, norm=emb)
-        return v.view(-1, 3, self.hid_dim)
-
-    def message(self, x_j, norm):
-        x_j = x_j.unsqueeze(1)
-        a = norm.view(-1, 3, self.hid_dim) * x_j
-        return a.view(-1, 3 * self.hid_dim)
-
-
-class EquiMessagePassing(MessagePassing):
-    def __init__(
-        self,
-        hidden_channels,
-        num_radial,
-    ):
-        super(EquiMessagePassing, self).__init__(aggr="add", node_dim=0)
-
-        self.hidden_channels = hidden_channels
-        self.num_radial = num_radial
-        self.dir_proj = nn.Sequential(
-            nn.Linear(3 * self.hidden_channels + self.num_radial, self.hidden_channels * 3),
-            nn.SiLU(inplace=True),
-            nn.Linear(self.hidden_channels * 3, self.hidden_channels * 3),
-        )
-
-        self.x_proj = nn.Sequential(
-            nn.Linear(hidden_channels, hidden_channels),
-            nn.SiLU(),
-            nn.Linear(hidden_channels, hidden_channels * 3),
-        )
-        self.rbf_proj = nn.Linear(num_radial, hidden_channels * 3)
-
-        self.inv_sqrt_3 = 1 / math.sqrt(3.0)
-        self.inv_sqrt_h = 1 / math.sqrt(hidden_channels)
-        self.x_layernorm = nn.LayerNorm(hidden_channels)
-
-        self.reset_parameters()
-
-    def reset_parameters(self):
-        nn.init.xavier_uniform_(self.x_proj[0].weight)
-        self.x_proj[0].bias.data.fill_(0)
-        nn.init.xavier_uniform_(self.x_proj[2].weight)
-        self.x_proj[2].bias.data.fill_(0)
-        nn.init.xavier_uniform_(self.rbf_proj.weight)
-        self.rbf_proj.bias.data.fill_(0)
-        self.x_layernorm.reset_parameters()
-        ## question: why don't reset parameters for dir_proj?
-
-    def forward(self, x, vec, edge_index, edge_rbf, weight, edge_vector):
-        xh = self.x_proj(self.x_layernorm(x))
-
-        rbfh = self.rbf_proj(edge_rbf)
-        weight = self.dir_proj(weight)
-        rbfh = rbfh * weight
-        # propagate_type: (xh: Tensor, vec: Tensor, rbfh_ij: Tensor, r_ij: Tensor)
-        dx, dvec = self.propagate(
-            edge_index,
-            xh=xh,
-            vec=vec,
-            rbfh_ij=rbfh,
-            r_ij=edge_vector,
-            size=None,
-        )
-
-        return dx, dvec
-
-    def message(self, xh_j, vec_j, rbfh_ij, r_ij):
-        x, xh2, xh3 = torch.split(xh_j * rbfh_ij, self.hidden_channels, dim=-1)
-        xh2 = xh2 * self.inv_sqrt_3
-
-        vec = vec_j * xh2.unsqueeze(1) + xh3.unsqueeze(1) * r_ij.unsqueeze(2)
-        vec = vec * self.inv_sqrt_h
-
-        return x, vec
-
-    def aggregate(
-        self,
-        features: Tuple[torch.Tensor, torch.Tensor],
-        index: torch.Tensor,
-        ptr: Optional[torch.Tensor],
-        dim_size: Optional[int],
-    ) -> Tuple[torch.Tensor, torch.Tensor]:
-        x, vec = features
-        x = scatter(x, index, dim=self.node_dim, dim_size=dim_size)
-        vec = scatter(vec, index, dim=self.node_dim, dim_size=dim_size)
-        return x, vec
-
-    def update(self, inputs: Tuple[torch.Tensor, torch.Tensor]) -> Tuple[torch.Tensor, torch.Tensor]:
-        return inputs
-
-
-class FTE(nn.Module):
-    def __init__(self, hidden_channels):
-        super().__init__()
-        self.hidden_channels = hidden_channels
-
-        self.vec_proj = nn.Linear(hidden_channels, hidden_channels * 2, bias=False)
-        self.xvec_proj = nn.Sequential(
-            nn.Linear(hidden_channels * 2, hidden_channels),
-            nn.SiLU(),
-            nn.Linear(hidden_channels, hidden_channels * 3),
-        )
-
-        self.inv_sqrt_2 = 1 / math.sqrt(2.0)
-        self.inv_sqrt_h = 1 / math.sqrt(hidden_channels)
-
-        self.reset_parameters()
-
-    def reset_parameters(self):
-        nn.init.xavier_uniform_(self.vec_proj.weight)
-        nn.init.xavier_uniform_(self.xvec_proj[0].weight)
-        self.xvec_proj[0].bias.data.fill_(0)
-        nn.init.xavier_uniform_(self.xvec_proj[2].weight)
-        self.xvec_proj[2].bias.data.fill_(0)
-
-    def forward(self, x, vec, node_frame):
-        vec = self.vec_proj(vec)
-        vec1, vec2 = torch.split(vec, self.hidden_channels, dim=-1)
-
-        # # # scalrization = torch.sum(vec1.unsqueeze(2) * node_frame.unsqueeze(-1), dim=1)
-        # # # scalrization[:, 1, :] = torch.abs(scalrization[:, 1, :].clone())
-        scalar = torch.sqrt(torch.sum(vec1**2, dim=-2) + 1e-10)
-
-        vec_dot = (vec1 * vec2).sum(dim=1)
-        vec_dot = vec_dot * self.inv_sqrt_h
-
-        x_vec_h = self.xvec_proj(torch.cat([x, scalar], dim=-1))
-        xvec1, xvec2, xvec3 = torch.split(x_vec_h, self.hidden_channels, dim=-1)
-
-        dx = xvec1 + xvec2 + vec_dot
-        dx = dx * self.inv_sqrt_2
-
-        dvec = xvec3.unsqueeze(1) * vec2
-
-        return dx, dvec
-
-
-class aggregate_pos(MessagePassing):
-    def __init__(self, aggr="mean"):
-        super(aggregate_pos, self).__init__(aggr=aggr)
-
-    def forward(self, vector, edge_index):
-        v = self.propagate(edge_index, x=vector)
-
-        return v
-
-
-class EquiOutput(nn.Module):
-    def __init__(self, hidden_channels):
-        super().__init__()
-        self.hidden_channels = hidden_channels
-
-        self.output_network = nn.ModuleList(
-            [
-                GatedEquivariantBlock(hidden_channels, 1),
-            ]
-        )
-
-        self.reset_parameters()
-
-    def reset_parameters(self):
-        for layer in self.output_network:
-            layer.reset_parameters()
-
-    def forward(self, x, vec):
-        for layer in self.output_network:
-            x, vec = layer(x, vec)
-        return vec.squeeze()
-
-
-# Borrowed from TorchMD-Net
-class GatedEquivariantBlock(nn.Module):
-    """Gated Equivariant Block as defined in Schütt et al. (2021):
-    Equivariant message passing for the prediction of tensorial properties and molecular spectra
-    """
-
-    def __init__(
-        self,
-        hidden_channels,
-        out_channels,
-    ):
-        super(GatedEquivariantBlock, self).__init__()
-        self.out_channels = out_channels
-
-        self.vec1_proj = nn.Linear(hidden_channels, hidden_channels, bias=False)
-        self.vec2_proj = nn.Linear(hidden_channels, out_channels, bias=False)
-
-        self.update_net = nn.Sequential(
-            nn.Linear(hidden_channels * 2, hidden_channels),
-            nn.SiLU(),
-            nn.Linear(hidden_channels, out_channels * 2),
-        )
-
-        self.act = nn.SiLU()
-
-    def reset_parameters(self):
-        nn.init.xavier_uniform_(self.vec1_proj.weight)
-        nn.init.xavier_uniform_(self.vec2_proj.weight)
-        nn.init.xavier_uniform_(self.update_net[0].weight)
-        self.update_net[0].bias.data.fill_(0)
-        nn.init.xavier_uniform_(self.update_net[2].weight)
-        self.update_net[2].bias.data.fill_(0)
-
-    def forward(self, x, v):
-        vec1 = torch.norm(self.vec1_proj(v), dim=-2)
-        vec2 = self.vec2_proj(v)
-
-        x = torch.cat([x, vec1], dim=-1)
-        x, v = torch.split(self.update_net(x), self.out_channels, dim=-1)
-        v = v.unsqueeze(1) * vec2
-
-        x = self.act(x)
-        return x, v
-
-
-class LEFTNetZ(nn.Module):
-    def __init__(
-        self,
-        atom_fea_dim,
-        num_targets,  # not used
-        otf_graph=False,
-        use_pbc=True,
-        regress_forces=False,
-        output_dim=0,
-        direct_forces=False,
-        cutoff=6.0,
-        num_layers=4,
-        readout="mean",
-        hidden_channels=128,
-        num_radial=96,
-        y_mean=0,
-        y_std=1,
-        eps=1e-10,
-    ):
-        super(LEFTNetZ, self).__init__()
-        self.y_std = y_std
-        self.y_mean = y_mean
-        self.eps = eps
-        self.num_layers = num_layers
-        self.hidden_channels = hidden_channels
-        self.cutoff = cutoff
-        self.num_targets = num_targets
-        self.readout = readout
-
-        self.regress_forces = regress_forces
-        self.use_pbc = use_pbc
-        self.otf_graph = otf_graph
-        self.direct_forces = direct_forces
-
-        self.input_dim = atom_fea_dim
-
-        # self.z_emb_ln = nn.LayerNorm(hidden_channels, elementwise_affine=False)
-        # self.z_emb = Embedding(95, hidden_channels)
-
-        self.radial_emb = rbf_emb(num_radial, cutoff)
-        self.radial_lin = nn.Sequential(
-            nn.Linear(num_radial, hidden_channels), nn.SiLU(inplace=True), nn.Linear(hidden_channels, hidden_channels)
-        )
-
-        self.neighbor_emb = NeighborEmb(self.hidden_channels, self.input_dim)
-
-        self.S_vector = S_vector(hidden_channels)
-
-        self.lin = nn.Sequential(
-            nn.Linear(3, hidden_channels // 4), nn.SiLU(inplace=True), nn.Linear(hidden_channels // 4, 1)
-        )
-
-        self.message_layers = nn.ModuleList()
-        self.FTEs = nn.ModuleList()
-
-        for _ in range(num_layers):
-            self.message_layers.append(EquiMessagePassing(hidden_channels, num_radial).jittable())
-            self.FTEs.append(FTE(hidden_channels))
-
-        if output_dim != 0:
-            self.num_targets = output_dim
-
-        self.last_layer = nn.Linear(hidden_channels, self.num_targets)
-        # self.out_forces = EquiOutput(hidden_channels)
-
-        # for node-wise frame
-        self.mean_neighbor_pos = aggregate_pos(aggr="mean")
-
-        self.inv_sqrt_2 = 1 / math.sqrt(2.0)
-
-        self.reset_parameters()
-
-    def reset_parameters(self):
-        # self.z_emb.reset_parameters()
-        self.radial_emb.reset_parameters()
-        for layer in self.message_layers:
-            layer.reset_parameters()
-        for layer in self.FTEs:
-            layer.reset_parameters()
-        self.last_layer.reset_parameters()
-        # self.out_forces.reset_parameters()
-        for layer in self.radial_lin:
-            if hasattr(layer, "reset_parameters"):
-                layer.reset_parameters()
-        for layer in self.lin:
-            if hasattr(layer, "reset_parameters"):
-                layer.reset_parameters()
-
-    def prop_setup(self, data, device):
-       '''
-       Setup for the property encoding. for LEFTNet-Z, only the atomic number is used as the input feature.
-       num_classes = 95 for the first 95 elements are used in the periodic table.
-       '''
-       z = data.atom_num.long().to(device)
-       z_encoded = F.one_hot(z, num_classes=95).float()
-       return z_encoded
-
-    # @conditional_grad(torch.enable_grad())
-    def _forward(self, data):
-        device = next(self.parameters()).device
-        pos = data.pos.to(device)
-        batch = data.batch.to(device)
-        z = self.prop_setup(data, device)
-
-
-        edge_index = radius_graph(pos, r=self.cutoff, batch=batch, max_num_neighbors=1000)
-        j, i = edge_index
-        vecs = pos[j] - pos[i]
-        dist = vecs.norm(dim=-1)
-
-        radial_emb = self.radial_emb(dist) # RBF, shape: (num_edges, num_radial=32), embedding for raw distance
-        radial_hidden = self.radial_lin(radial_emb) # MLP
-        rbounds = 0.5 * (torch.cos(dist * pi / self.cutoff) + 1.0) # for soft cutoff, smooth decay, value closer to 1 means stronger relationship, closer to 0 means weaker
-
-        radial_emb = radial_emb.to(device)
-        radial_hidden = radial_hidden.to(device)
-        rbounds = rbounds.to(device)
-
-        radial_hidden = rbounds.unsqueeze(-1) * radial_hidden # further stengthen the representation of radial_emb
-
-        # init invariant node features
-        # shape: (num_nodes, hidden_channels)
-        s = self.neighbor_emb(z, edge_index, radial_hidden) # z (num_nodes, atom_encoding), z_emb (num_nodes, hidden_channels); s=z_emb+neighbor_emb, shape: (num_nodes, hidden_channels)
-
-        # init equivariant node features
-        # shape: (num_nodes, 3, hidden_channels)
-        vec = torch.zeros(s.size(0), 3, s.size(1), device=s.device)
-
-        # bulid edge-wise frame
-        edge_diff = vecs
-        edge_diff = edge_diff / (dist.unsqueeze(1) + self.eps) # normalize the edge_diff
-        # noise = torch.clip(torch.empty(1,3).to(z.device).normal_(mean=0.0, std=0.1), min=-0.1, max=0.1)
-        edge_cross = torch.cross(pos[i], pos[j])
-        edge_cross = edge_cross / ((torch.sqrt(torch.sum((edge_cross) ** 2, 1).unsqueeze(1))) + self.eps)
-        edge_vertical = torch.cross(edge_diff, edge_cross)
-        # shape: (num_edges, 3, 3)
-        edge_frame = torch.cat((edge_diff.unsqueeze(-1), edge_cross.unsqueeze(-1), edge_vertical.unsqueeze(-1)), dim=-1) # normalised edge, edge_cross, edge_vertical
-
-        node_frame = 0
-
-        # LSE: local 3D substructure encoding
-        # S_i_j shape: (num_nodes, 3, hidden_channels)
-        S_i_j = self.S_vector(s, edge_diff.unsqueeze(-1), edge_index, radial_hidden)
-        scalrization1 = torch.sum(S_i_j[i].unsqueeze(2) * edge_frame.unsqueeze(-1), dim=1)
-        scalrization2 = torch.sum(S_i_j[j].unsqueeze(2) * edge_frame.unsqueeze(-1), dim=1)
-        scalrization1[:, 1, :] = torch.abs(scalrization1[:, 1, :].clone())
-        scalrization2[:, 1, :] = torch.abs(scalrization2[:, 1, :].clone())
-
-        scalar3 = (
-            self.lin(torch.permute(scalrization1, (0, 2, 1)))
-            + torch.permute(scalrization1, (0, 2, 1))[:, :, 0].unsqueeze(2)
-        ).squeeze(-1)
-        scalar4 = (
-            self.lin(torch.permute(scalrization2, (0, 2, 1)))
-            + torch.permute(scalrization2, (0, 2, 1))[:, :, 0].unsqueeze(2)
-        ).squeeze(-1)
-
-        edge_weight = torch.cat((scalar3, scalar4), dim=-1) * rbounds.unsqueeze(-1)
-        edge_weight = torch.cat((edge_weight, radial_hidden, radial_emb), dim=-1)
-
-        # for i in range(self.num_layers):
-        for i in range(1):
-            ds, dvec = self.message_layers[i](s, vec, edge_index, radial_emb, edge_weight, edge_diff)
-
-            s = s + ds
-            vec = vec + dvec
-
-            # FTE: frame transition encoding
-            ds, dvec = self.FTEs[i](s, vec, node_frame)
-
-            s = s + ds
-            vec = vec + dvec
-
-        s = self.last_layer(s)
-        s = scatter(s, batch, dim=0, reduce=self.readout)
-        s = s * self.y_std + self.y_mean
-        return s
-
-    def forward(self, data):
-        return self._forward(data)
-
-    @property
-    def num_params(self):
-        return sum(p.numel() for p in self.parameters())
-
-
-
-class LEFTNetProp(LEFTNetZ):
-    '''
-    LEFTNet with property encoding, using a one-hot encoding of the atomic properties for the atoms.
-    Rest of the architecture is the same as LEFTNetZ.
-    '''
-    def __init__(
-        self,
-        atom_fea_dim,
-        num_targets,  # not used
-        otf_graph=False,
-        use_pbc=True,
-        regress_forces=False,
-        output_dim=0,
-        direct_forces=False,
-        cutoff=6.0,
-        num_layers=4,
-        readout="mean",
-        hidden_channels=128,
-        num_radial=96,
-        y_mean=0,
-        y_std=1,
-        eps=1e-10,
-    ):
-        super(LEFTNetProp, self).__init__(
-            atom_fea_dim=atom_fea_dim,
-            num_targets=num_targets,
-            otf_graph=otf_graph,
-            use_pbc=use_pbc,
-            regress_forces=regress_forces,
-            output_dim=output_dim,
-            direct_forces=direct_forces,
-            cutoff=cutoff,
-            num_layers=num_layers,
-            readout=readout,
-            hidden_channels=hidden_channels,
-            num_radial=num_radial,
-            y_mean=y_mean,
-            y_std=y_std,
-            eps=eps
-        )
-    def prop_setup(self, data, device):
-       '''
-       Setup for the property encoding. for LEFTNet-Prop, a one-hot encoding contains atomic properties is used as the atomic feature..
-       '''
-
-       return data.x.to(device)
