--- examples/materials_benchmark/move_to_kale/prepdata/materials_transform.py
@@ -1,45 +1,0 @@
-import numpy as np
-
-class GaussianDistance(object):
-    """
-    Expands the distance by Gaussian basis.
-
-    Unit: angstrom
-    """
-    def __init__(self, dmin, dmax, step, var=None):
-        """
-        Parameters
-        ----------
-
-        dmin: float
-          Minimum interatomic distance
-        dmax: float
-          Maximum interatomic distance
-        step: float
-          Step size for the Gaussian filter
-        """
-        assert dmin < dmax
-        assert dmax - dmin > step
-        self.filter = np.arange(dmin, dmax+step, step)
-        if var is None:
-            var = step
-        self.var = var
-
-    def expand(self, distances):
-        """
-        Apply Gaussian disntance filter to a numpy distance array
-
-        Parameters
-        ----------
-
-        distance: np.array shape n-d array
-          A distance matrix of any shape
-
-        Returns
-        -------
-        expanded_distance: shape (n+1)-d array
-          Expanded distance matrix with the last dimension of length
-          len(self.filter)
-        """
-        return np.exp(-(distances[..., np.newaxis] - self.filter)**2 /
-                      self.var**2)
